requirements.txt
asgiref==3.5.1
boto3==1.24.17
botocore==1.27.17
cycler==0.11.0
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-storages==1.12.3
djangorestframework==3.13.1
fonttools==4.33.3
gunicorn==20.1.0
jmespath==1.0.1
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
